Updated to create /etc/openshift/htpasswd if it doesn't exist
Wasn't creating it with batch only flag set because it was attempting to batch add on a non-existent file.  Adding the -c creates the file if it doesn't exist, and still allows for the batch processing to occur.  (-b in this case accepts it from stdout of the .pp instead of from a file.)