Fix unwanted space in front of MONGO_HOST_PORT line
There is a typo (white space) that was introduced unintentionally into
the code base the previous commit:

<https://github.com/openshift/puppet-openshift_origin/commit/22e3bbbab10133fef3d4b414030cc6833bc42363>

The white space occupied right before MONGO_HOST_PORT line. This unwanted
issue cause the oo-accept-broker to fail in Enterprise devenv as it
couldn't locate the MONGO_HOST_PORT line.

This commit will remove the space and the oo-accept-broker shoyld be able
to locate the line for mango host info correctly.

Signed-off-by: Vu Dinh <vdinh@redhat.com>